Add files via upload
This is a reupload of some exploratory analysis performed on mock data related to the UK National Rail. The file had been updated to include the application/tool used to create the file. After this file has been uploaded, the previous version of this file in the "general data analytics samples" repository fill be deleted.