Add files via upload
This is a reupload of a file that features some exploratory analysis and an exploratory dashboard that use mock data related to the UK National Rail. The file name had been updated to include the application/tool used to create this file. After this file has been uploaded to the repository dedicated to the UK National Rail project, the previous file in the "general data analytics samples" repository will be deleted.