Add files via upload
This is a reupload of a file that contains an exploratory dashboard made with Power BI Desktop using mock data related to the UK National Rail. After this file has been uploaded to a repository dedicated to the UK National Rail project, the previous version of this file in the "general data analytics samples" repository will be deleted.