Create UK_National_Rail_MOCK_DATA_README
This is a recreation of a README file created to help explain the UK National Rail Project. The file name had been updated to emphasize the fact that mock data had been used for this project. After this file has been created, the previous README file in the "general data analytics samples" repository will be deleted.